Move components into Sidenav folder
