+++ src/components/Navbar.js
@@ -4,8 +4,7 @@ import { css } from '@emotion/core';
 import { Logo } from '@leafygreen-ui/logo';
 import { uiColors } from '@leafygreen-ui/palette';
 import Searchbar from './Searchbar';
-import { SidenavContext } from './sidenav-context';
-import SidenavMobileMenuButton from './SidenavMobileMenuButton';
+import { SidenavContext, SidenavMobileMenuButton } from './Sidenav';
 import useMedia from '../hooks/use-media';
 import { theme } from '../theme/docsTheme';
 import { getSearchbarResultsFromJSON } from '../utils/get-searchbar-results-from-json';
