+++ src/components/RootProvider.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { HeaderContextProvider } from './header-context';
 import { NavigationProvider } from './navigation-context';
-import { SidenavContextProvider } from './sidenav-context';
+import { SidenavContextProvider } from './Sidenav';
 import { TabProvider } from './tab-context';
 
 const RootProvider = ({ children, isSidenavEnabled, selectors }) => (
