+++ src/components/Sidenav/index.js
@@ -1,0 +1,6 @@
+import Sidenav from './Sidenav';
+import SidenavBackButton from './SidenavBackButton';
+import SidenavMobileMenuButton from './SidenavMobileMenuButton';
+import { SidenavContext, SidenavContextProvider } from './sidenav-context';
+
+export { Sidenav, SidenavBackButton, SidenavContext, SidenavContextProvider, SidenavMobileMenuButton };
